vendors

+++ resources/react/products/src/components/Categories/CategoryItem.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import {Button,Image} from 'react-bootstrap'
+import { Button, Image } from 'react-bootstrap'
 
 class CategoryItem extends React.Component {
     constructor(props) {
@@ -9,8 +9,14 @@ class CategoryItem extends React.Component {
             item: props.item
         }
 
+        this.edit = () => {
+            if ('function' == typeof (props.onEdit))
+                props.onEdit(this.state.item)
+        }
+
         this.delete = () => {
-            props.onDelete(this.state.item)
+            if ('function' == typeof (props.onDelete))
+                props.onDelete(this.state.item)
         }
     }
 
@@ -20,9 +26,9 @@ class CategoryItem extends React.Component {
         return <>
             <tr id={"cat-" + item.id}>
                 <td>{item.id}</td>
-                <td>{item.name} {item.parents?(<><br /><small>{item.parents.map((e)=>{return e.name+" / "})}</small></>):''}</td>
-                <td><div className="preview"><Image src={(item.image && item.image.indexOf('://')>-1)?item.image:ASSETS_URL+item.image}></Image></div></td>
-                <td><Button variant="success" href={"#/edit-category/"+item.id}><i className="fas fa-pencil-alt"></i> Edit</Button> <Button variant="danger" onClick={this.delete}><i className="fas fa-trash-alt"></i> Delete</Button> </td>
+                <td>{item.name} {item.parents ? (<><br /><small>{item.parents.map((e) => { return e.name + " / " })}</small></>) : ''}</td>
+                <td><div className="preview"><Image src={(item.image && item.image.indexOf('://') > -1) ? item.image : ASSETS_URL + item.image}></Image></div></td>
+                <td><Button variant="success" href={"#/edit-category/" + item.id}><i className="fas fa-pencil-alt"></i> Edit</Button> <Button variant="danger" onClick={this.delete}><i className="fas fa-trash-alt"></i> Delete</Button> </td>
             </tr>
         </>
     }
